+++ packages/core/transpiler/reactivity-parser/src/types.ts
@@ -67,6 +67,18 @@ export interface IfParticle {
   branches: IfBranch[]
 }
 
+export interface SwitchBranch {
+  case: DependencyValue<t.Expression> | null | undefined
+  children: ViewParticle[]
+  break: boolean
+}
+
+export interface SwitchParticle {
+  type: "switch"
+  discriminant: DependencyValue<t.Expression>
+  branches: SwitchBranch[]
+}
+
 export interface EnvParticle {
   type: "env"
   props: Record<string, DependencyProp>
@@ -95,6 +107,7 @@ export type ViewParticle =
   | IfParticle
   | EnvParticle
   | ExpParticle
+  | SwitchParticle
   | SubviewParticle
 
 export interface ReactivityParserConfig {