+++ packages/core/transpiler/reactivity-parser/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @dlightjs/reactivity-parser
 
+## 1.0.0-alpha.2
+
+### Patch Changes
+
+- feat: add switch statement support
+- Updated dependencies
+  - @dlightjs/view-parser@1.0.0-alpha.1
+
 ## 1.0.0-alpha.1
 
 ### Patch Changes